refactor: utils 패키지 제거
