+++ src/main/java/lotto/domain/analysis/Analysis.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 import lotto.domain.money.Money;
-import lotto.utils.Rank;
+import lotto.domain.rank.Rank;
 
 public class Analysis {
