+++ src/main/java/lotto/domain/ball/condition/BallNumberRange.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.domain.ball.condition;
 
 import java.util.stream.IntStream;
