+++ src/main/java/lotto/domain/money/condition/MoneyUnit.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.domain.money.condition;
 
 public enum MoneyUnit {
