+++ src/main/java/lotto/domain/rank/Rank.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.domain.rank;
 
 import java.util.Arrays;
 import java.util.List;
