+++ src/main/java/lotto/domain/ticket/condition/TicketSize.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.domain.ticket.condition;
 
 public enum TicketSize {
