+++ src/main/java/lotto/domain/winning/WinningTicket.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 
 import lotto.domain.ball.Ball;
 import lotto.domain.ball.BallStorage;
+import lotto.domain.rank.Rank;
 import lotto.domain.ticket.Ticket;
-import lotto.utils.Rank;
 
 public class WinningTicket {
