+++ src/main/java/lotto/view/output/OutputView.java
@@ -10,10 +10,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import lotto.domain.rank.Rank;
 import lotto.dto.AnalysisDto;
 import lotto.dto.TicketBundlesDto;
 import lotto.dto.TicketDto;
-import lotto.utils.Rank;
 import lotto.view.utils.Delimiter;
 
 public class OutputView {
