+++ src/test/java/lotto/domain/ticket/TicketBundlesTest.java
@@ -8,9 +8,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import lotto.domain.rank.Rank;
 import lotto.domain.ticket.generator.CustomTicketGenerator;
 import lotto.domain.winning.WinningTicket;
-import lotto.utils.Rank;
 
 class TicketBundlesTest {
