+++ src/test/java/lotto/domain/winning/WinningTicketTest.java
@@ -11,10 +11,10 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import lotto.domain.rank.Rank;
 import lotto.domain.ticket.Ticket;
 import lotto.exception.LottoException;
 import lotto.exception.LottoExceptionStatus;
-import lotto.utils.Rank;
 
 class WinningTicketTest {
